+++ pysys-examples/internal/utilities/resources/runpysys.py
@@ -9,7 +9,9 @@ def runPySys(processowner, stdouterr, args, ignoreExitStatus=False, abortOnError
 		args = ['-m', 'pysys']+args
 	env = dict(environs or {})
 	for k in os.environ: 
-		if k not in env and not k.startswith('PYSYS_'): env[k] = os.environ[k]
+		# don't preserve any env vars from parent env that might affect test behaviour
+		if k not in env and not k.startswith('PYSYS_') and \
+			k not in ['TRAVIS']: env[k] = os.environ[k]
 	for k in list(env.keys()):
 		if env[k] == None: env.pop(k)
 	pypath = os.path.dirname(sys.executable)
@@ -25,9 +27,12 @@ def runPySys(processowner, stdouterr, args, ignoreExitStatus=False, abortOnError
 		# output for child processes; otherwise we'd only get coverage for things 
 		# that can be tested using the top-level pysys invocation
 		args = ['-m', 'coverage', 'run', '--parallel-mode']+args
-	return processowner.startProcess(command=sys.executable,
-		arguments = args,
-		environs = env, ignoreExitStatus=ignoreExitStatus, abortOnError=abortOnError, 
-		stdout=stdouterr+'.out', stderr=stdouterr+'.err', 
-		displayName='pysys %s'%stdouterr, 
-		**kwargs)
+	try:
+		return processowner.startProcess(command=sys.executable,
+			arguments = args,
+			environs = env, ignoreExitStatus=ignoreExitStatus, abortOnError=abortOnError, 
+			stdout=stdouterr+'.out', stderr=stdouterr+'.err', 
+			displayName='pysys %s'%stdouterr, 
+			**kwargs)
+	finally: # in case there was any error printed to stderr
+		processowner.logFileContents(stdouterr+'.err')