+++ CHANGELOG.rst
@@ -85,10 +85,26 @@ New features
   outcomes were logged (so if you have a complex test you can see at a glance 
   if there's just one problem to resolve, or 5, or 20!).
 
-- Added `runLogOutput=` parameter to the `processResult()` method of 
-  the `BaseResultsWriter` API so that writers such as the 
-  `JUnitXMLResultsWriter` can include the test output with no loss of unicode 
-  character information. 
+- Extended the writers API:
+   - Added `runLogOutput=` parameter to the `processResult()` method of 
+     the `BaseResultsWriter` class so that writers such as the 
+     `JUnitXMLResultsWriter` can include the test output with no loss of unicode 
+     character information. 
+   - Added `testoutdir=` parameter to the `setup()` method so writers have 
+     a way to identify different test runs on the same machine. 
+   - Added `runner=` parameter to the `setup()` method so writers have 
+     access to the runner instance for reading/modifying configuration 
+     settings. 
+   - Added `isEnabled()` method that can optionally be used by a writer to 
+     disable itself based on the environment in which it is running, or 
+     to enable itself even when `--record` isn't specified, which is useful 
+     for writers that produce output for a CI system. 
+
+- Added support for running PySys tests under Travis CI to the sample 
+  pysysproject.xml file, including by default only printing run.log output 
+  for failed tests, and containing that detailed output within a folded 
+  section that can be expanded if needed. To enable this just ensure that the 
+  Travis CI writer is enabled in your project configuration file. 
 
 - Added support for configuring the default encodings to use for common file 
   patterns in the `pysysproject.xml` configuration, e.g. ::
@@ -111,6 +127,18 @@ New features
   stored in `run.log`. PySys will now catch output written using `print()` 
   statements and redirect it to the logging framework, so it will show up 
   in `run.log`. 
+
+- Added new command line option `--printLogs all|failures|none` (default value 
+  is `all`) which allows user to avoid the printing of run.log to the stdout 
+  console either for all tests, or for tests that pass. This is useful to 
+  avoid generating huge amounts of output during large test runs (which can 
+  be problematic when stdout is captured by a Continuous Integration system), 
+  or to show detailed information only for failing tests which makes it easier 
+  for a user to locate the diagnostic information they care about more quickly. 
+  The `printLogs` option is stored in runner.printLogs` and can be changed by 
+  custom writer implementations if desired, for example to avoid duplicating 
+  information already being printed to stdout by the writer in a different 
+  format. 
 
 - As an alternative to the usual `pysys.py` executable script, it is now also 
   possible to launch PySys using::